mettler: get_mass_stable now averages three readings
Also when reading the balance, it checks if the balance has been read correctly by ensuring that the unit is either as specified in the Balance Register, or is one of the other units in SUFFIX (in which case it resets the unit in the software instance to the actual unit of the balance). Otherwise it complains that either there was a communication error or the balance unit is set incorrectly.