All levels of pop-ups work; scheme table allows drag drop of Excel file
This upgrade uses a new 'PromptThread' with its own 'PromptWorker' to control the user-interaction popups for data collection.  These bits may need to appear on another computer at a later date (e.g. turn the circweigh_popup into a service).

Also added config to git as example, and changed the standards to read to 1 s.f.