Modifications to ambient monitoring setup
Specific omega sensor information is vital as the different sensors are attached to different balances.  Added a column in balance register to include which omega and sensor (or Vaisala unit) is used for a given balance.
Also, ambient monitoring info is now saved as part of each measurement run metadata.