Added Signal-Slot to update balance list from config file
When the operator 'confirms housekeeping info', which triggers initialisation, the equipment database (from config file) is searched for available balances and these are transferred to the scheme table.  'Folder' in config file is also changed to the parent folder, assuming that the subfolder in which all data is saved uses the name of the client anyway.