changes to at106.py as used for recent calibration
adjustments to electronic weights procedure to only load or unload electronic weights when a mass is loaded. made ask_internal_weights into a method. also modified the configuration options to include a few more initialisation commands, and moved this to the check loading routine to be called when a mass is loaded.