added test cases from TP MSLT.M.001.008 appendices C and D
Test cases are for a single circular weighing and a matrix least squares analysis for a set of 9 masses respectively.
To use a relative uncertainty of 0.1 ppm for no buoyancy correction, a change was made in the FinalMassCalc class to include REL_UNC as a class variable so that the value could be changed in this class instance only.