allow completion without ambient monitoring for aw mode
To enable this change, checks have been added both before and after a set of automatic weighings to ensure that some ambient data is recorded. If no data was recorded for any weighing in a scheme entry, these check values can be manually entered into the summary.xlsx file when it is exported via a pop-up (which can also be safely ignored). The balance mode is used to ensure only automatic weighings are allowed to proceed without ambient monitoring. The number of attempts at contacting the OMEGA server has been set at 10 for both 'now' and 'during' requests. The summary outputs have been adapted to cope with missing values.