Added direct read of chamber monitoring databases for BuildDown2023
To use this change, specify 'BuildDown' in the Ambient monitoring column in the Balance Register. Ensure that the database directory 'database_dir' is correct in constants.py, and contains the databases with names 'Temperature_milliK.sqlite3' and 'TRHP_Vaisala.sqlite3'. Changes to the database reading may be necessary when running Mohsen's code for adjusting the pressure of the chamber.
Note that the calibration information for the milliK and PRT is hard-coded in ambient_fromdatabase.py and will need entering/checking for each equipment set-up.